LayoutTests/platform/wpe/imported/w3c/web-platform-tests/encrypted-media/clearkey-mp4-update-disallowed-input-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.

Harness Error (TIMEOUT), message = null

TIMEOUT update() with invalid response (longer than 64Kb characters) should fail. Test timed out
